added texture system, changed names of structs and file, reorganized repo

+++ src/resources/resource_types.rs
@@ -1,4 +1,4 @@
-use super::super::vulkan::vulkan_image::VulkanImage;
+use crate::application::renderer::vulkan::vulkan_image::VulkanImage;
 use ash::vk::Sampler;
 
 #[derive(Clone, Copy)]
@@ -8,7 +8,7 @@ pub struct Texture {
     pub height: u32,
     pub channel_count: u8,
     pub has_transparency: bool,
-    pub generation: u32,
+    pub generation: usize,
     pub internal_data: TextureData,
 }
 #[derive(Clone, Copy)]
